# packages/contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: aos-api
  version: 0.3.0
  description: >
    Stable contract surface per docs/palantier/20_tech/T-API-aos-api稳定契约.md.
    Aligned Wave-0～5 + G-ALIGN (47) — main paths listed; see T-API for full narrative.
servers:
  - url: /v1
paths:
...
  /ontology/object-types:
    get:
      operationId: listObjectTypes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ok
    post:
      operationId: createObjectType
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ok
  /ontology/link-types:
    get:
      operationId: listLinkTypes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ok
    post:
      operationId: createLinkType
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ok
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...